multiple fixes
